move Stripe effect into reusable hook
