+++ src/components/account-settings-overlay/subscription-settings.js
@@ -3,22 +3,12 @@ import { Button } from '@fogcreek/shared-components';
 
 import Heading from 'Components/text/heading';
 import Text from 'Components/text/text';
-import useScript from 'Hooks/use-script';
+import useStripe from 'State/stripe';
 
 function SubscriptionSettings() {
   // Use temporary state variable to toggle subscription status
   const [subscribed, setSubscribed] = useState(false);
-  const [stripe, setStripe] = useState(null);
-
-  useScript('https://js.stripe.com/v3/');
-
-  useEffect(() => {
-    const initStripe = () => {
-      setStripe(window.Stripe('pk_test_WIJxKl0T1xT8zSb2StHu8zlo'));
-    };
-    const script = document.querySelector('script[src="https://js.stripe.com/v3/"]');
-    script.addEventListener('load', initStripe);
-  });
+  const stripe = useStripe();
 
   return (
     <>
